elt_doc_assessment/pyproject.toml
[project]
name = "elt-doc-assessment"
version = "0.1.0"
description = "Assessment tool for ETL/ELT document URLs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "python-docx>=1.2.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.2",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "Pillow>=10.0.0",
    "httpx>=0.25.0",
    "click>=8.0.0",
]

[project.scripts]
elt-doc-assessment = "elt_doc_assessment.cli:cli"

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=4.0.0",
]
